Add support for hash-based RUG scale out
This adds support for running multiple RUG processes to scale out
and distribute appliance load across them.  It uses a hash ring implementation
lifted from Ironic (with modification).  The gist is:

* Workers now maintain a copy of the hash ring, which is hashed using the
list of members in the cluster.

* A new subprocess connects to an external coordination service via tooz, ie
memcache or zookeeper.  This service's only purpose is to track cluster
membership and report changes to this subprocess.  On membership changes,
the coordination subprocess creates a REBALANCE event and puts it on the
internal notification queue.  There is no leadership election required.

* When a worker gets a REBALANCE event, it rebalances the hash ring based
on the new membership list.

* Prior to processing any events bound for a specified router, the worker
first checks the hash manager to find if the resource is assigned to it.
If not, it ignores it.  If it is, it processes the event. This also applies
to incoming command events.

Partially implements: blueprint rug-scaling

+++ akanda/rug/main.py
@@ -28,6 +28,7 @@ from oslo_log import log
 
 from akanda.rug.common.i18n import _, _LE, _LI
 from akanda.rug.common import config as ak_cfg
+from akanda.rug import coordination
 from akanda.rug import daemon
 from akanda.rug import health
 from akanda.rug import metadata
@@ -148,6 +149,18 @@ def main(argv=sys.argv[1:]):
     )
     notification_proc.start()
 
+    if CONF.coordination.enabled:
+        coordinator_proc = multiprocessing.Process(
+            target=coordination.start,
+            kwargs={
+                'notification_queue': notification_queue
+            },
+            name='coordinator',
+        )
+        coordinator_proc.start()
+    else:
+        coordinator_proc = None
+
     mgt_ip_address = neutron_api.get_local_service_ip(cfg.CONF).split('/')[0]
     metadata_proc = multiprocessing.Process(
         target=metadata.serve,
@@ -201,6 +214,9 @@ def main(argv=sys.argv[1:]):
         publisher.stop()
 
         # Terminate the subprocesses
-        for subproc in [notification_proc, metadata_proc, rug_api_proc]:
+        for subproc in [notification_proc, coordinator_proc, metadata_proc,
+                        rug_api_proc]:
+            if not subproc:
+                continue
             LOG.debug(_('Stopping %s.'), subproc.name)
             subproc.terminate()
